Modifica los controladores de edición de un producto. Para que eliminen las fotos de los mismos solo si existe el archivo en el proyecto. Para evitar errores de eliminación manual de la imágen y luego querer eliminar el producto utilizando el controlador.

+++ src/controllers/activitiesControllers.js
@@ -167,7 +167,9 @@ module.exports = {
                         `../../public/${image.image}`
                     );
                     // lo elimina
-                    await fs.promises.unlink(imagePath);
+                    if (fs.existsSync(imagePath)) {
+                        await fs.promises.unlink(imagePath);
+                    }
                     await image.destroy();
                 }
             }
@@ -246,7 +248,9 @@ module.exports = {
                         `../../public/${image.image}`
                     );
                     // lo elimina
-                    await fs.promises.unlink(imagePath);
+                    if (fs.existsSync(imagePath)) {
+                        await fs.promises.unlink(imagePath);
+                    }
                     await image.destroy();
                 }
             }
