+++ src/controllers/cottagesController.js
@@ -130,7 +130,9 @@ module.exports = {
                         `../../public/${image.image}`
                     );
                     // lo elimina
-                    await fs.promises.unlink(imagePath);
+                    if (fs.existsSync(imagePath)) {
+                        await fs.promises.unlink(imagePath);
+                    }
                     await image.destroy();
                 }
             }
@@ -204,7 +206,10 @@ module.exports = {
                         `../../public/${image.image}`
                     );
                     // lo elimina
-                    await fs.promises.unlink(imagePath);
+                    if (fs.existsSync(imagePath)) {
+                        await fs.promises.unlink(imagePath);
+                    }
+
                     await image.destroy();
                 }
             }
